Move common variables to a common file in db package
