+++ go-backend/service/user_http_test.go
@@ -60,20 +60,3 @@ func (suite *UsersHandlerTestSuite) TestListUsersWhenDBFailure() {
 	suite.dbMock.AssertExpectations(suite.T())
 }
 
-// path: is used to configure router path (eg: /users/{id})
-// requestURL: current request path (eg: /users/1)
-func makeHTTPCall(method, path, requestURL, body string, handlerFunc http.HandlerFunc) (recorder *httptest.ResponseRecorder) {
-	// create a http request using the given parameters
-	req, _ := http.NewRequest(method, requestURL, strings.NewReader(body))
-
-	// test recorder created for capturing api responses
-	recorder = httptest.NewRecorder()
-
-	// create a router to serve the handler in test with the prepared request
-	router := mux.NewRouter()
-	router.HandleFunc(path, handlerFunc).Methods(method)
-
-	// serve the request and write the response to recorder
-	router.ServeHTTP(recorder, req)
-	return
-}
